test ImageCommentsMapper in isolation
